+++ CHANGELOG.md
@@ -53,6 +53,11 @@ different versioning scheme, following the Haskell community's
       and
       [the bf example](https://github.com/Microsoft/bond/commit/11beaf5319639e4bdee96a25f95154e4fed93a75#diff-bdda0f39d99280d4858b4453906eea17)
       for more details.
+* **Breaking change** The `bond::customize<protocols>` has been removed. All the
+  public APIs that require a protocol list (e.g. `bond::Marshal`) now accept
+  an extra template argument `Protocols` which defaults to `bond::BuiltInProtocols`.
+  Custom input streams now require `bond::type_id<>` to be specialized with a
+  unique magic number. For more details please see [the bf example](https://github.com/Microsoft/bond/tree/master/examples/cpp/core/bf).
 * Initial support for sending
   [Bond objects over gRPC](https://microsoft.github.io/bond/manual/bond_over_grpc.html)
   has been added.